computer_genie/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "computer-genie"
version = "1.0.0"
description = "AI-powered computer automation framework"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Computer Genie Team", email = "team@computergenie.ai"}
]
keywords = ["automation", "ai", "rpa", "computer-vision", "ui-automation"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.24.0",
    "websockets>=11.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.2",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "anthropic>=0.20.0",
    "openai>=1.0.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "easyocr>=1.7.0",
    "mss>=9.0.0",
    "pynput>=1.7.6",
    "psutil>=5.9.0",
    "aiofiles>=23.0.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "prometheus-client>=0.17.0",
    "structlog>=23.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]
chat = [
    "streamlit>=1.25.0",
    "gradio>=3.40.0",
]
enterprise = [
    "ldap3>=2.9.0",
    "cryptography>=41.0.0",
    "pyjwt>=2.8.0",
]

[project.scripts]
genie = "computer_genie.cli.main:app"

[project.urls]
Homepage = "https://github.com/yourusername/computer-genie"
Documentation = "https://computer-genie.readthedocs.io"
Repository = "https://github.com/yourusername/computer-genie"
Issues = "https://github.com/yourusername/computer-genie/issues"